Rename MedicationRequest.ts to medicationRequest.ts
